+++ browser/cli/src/generateBaseObject.ts
@@ -1,6 +1,7 @@
-import { core, Resource } from '@tomic/lib';
+import { Resource, type Core } from '@tomic/lib';
 import { store } from './store.js';
 import { camelCaseify, dedupe } from './utils.js';
+import chalk from 'chalk';
 
 export type ReverseMapping = Record<string, string>;
 
@@ -10,21 +11,19 @@ type BaseObject = {
 };
 
 export const generateBaseObject = async (
-  ontology: Resource,
+  ontology: Resource<Core.Ontology>,
 ): Promise<[string, ReverseMapping]> => {
   if (ontology.error) {
     throw ontology.error;
   }
 
-  const classes = dedupe(ontology.get(core.properties.classes)) as string[];
-  const properties = dedupe(
-    ontology.get(core.properties.properties),
-  ) as string[];
+  const classes = dedupe(ontology.props.classes ?? []);
+  const properties = dedupe(ontology.props.properties ?? []);
   const name = camelCaseify(ontology.title);
 
   const baseObj = {
-    classes: await listToObj(classes),
-    properties: await listToObj(properties),
+    classes: await listToObj(classes, 'classes'),
+    properties: await listToObj(properties, 'properties'),
   };
 
   const objStr = `export const ${name} = {
@@ -35,7 +34,10 @@ export const generateBaseObject = async (
   return [objStr, createReverseMapping(name, baseObj)];
 };
 
-const listToObj = async (list: string[]): Promise<Record<string, string>> => {
+const listToObj = async (
+  list: string[],
+  type: string,
+): Promise<Record<string, string>> => {
   const entries = await Promise.all(
     list.map(async subject => {
       const resource = await store.getResource(subject);
@@ -43,6 +45,28 @@ const listToObj = async (list: string[]): Promise<Record<string, string>> => {
       return [camelCaseify(resource.title), subject];
     }),
   );
+
+  // check for duplicates and throw an error if there are any.
+  const duplicates = entries.filter(
+    (entry, index) => entries.findIndex(e => e[0] === entry[0]) !== index,
+  );
+
+  if (duplicates.length > 0) {
+    // eslint-disable-next-line no-console
+    console.log(
+      chalk.red(`ERROR: Found ${type} with the same name: `),
+      duplicates.map(e => e[0]).join(', '),
+    );
+
+    // eslint-disable-next-line no-console
+    console.log(
+      chalk.red(
+        'Properties with the same name will conflict in the generated ontology. Try to reuse properties where possible or rename the duplicate to prevent a conflict.',
+      ),
+    );
+
+    process.exit(1);
+  }
 
   return Object.fromEntries(entries);
 };